third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-item-end-after-get-set-expected.txt

Test that setting and getting grid-end and grid-after works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-end and grid-after set through CSS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('grid-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-after') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto / auto'

Test getting and setting grid-end and grid-after through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue

Test setting grid-start and grid-before to 'inherit' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue

Test setting grid-start and grid-before to 'initial' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-after') is gridAfterValue

Test setting grid-end and grid-after back to 'auto' through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-end') is '18'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto / 18'
PASS getComputedStyle(element, '').getPropertyValue('grid-after') is '66'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto / 66'
PASS getComputedStyle(element, '').getPropertyValue('grid-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-after') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto / auto'
PASS successfullyParsed is true

TEST COMPLETE
